feat: only throw error when subscribed to task instance as promise
Now, performing a task will only throw an error if the resulting promise has been subscribed to (meaning the promise is being monitored with `.then` or `.catch`, or it was used with `async`/`await`).

This allows you to perform a task and handle the error declaritively, using the `.error` property on the `TaskInstance` where it makes sense, but still handle it imperitively using a more traditional approach if desired.