refactor: move task state to `current` property
Inspired by React Refs and how they can be used to model mutable state
within React, I have decided that it might be easier to organize the
current state of a task in the same way, as it too can change over time.

This change allows a `TaskInstance` to be returned immediately after
performing it, but still allows for handling a change in state the same
way that one would handle any other change to a Ref.

Now, properties about the state of a task will be accessed off of
`TaskInstance.current`.

The `.lastSuccessful` property was updated to be an instance of `TaskState`,
since once a task has completed, the ability to await the result or cancel
the task is irrelevant.