add maxConcurrent option
This uses the existing instances list as a queue. Instead of checking
derivedState.isRunning for keep: "first" and keep: "last", it compares
the number of running instances against maxConcurrency.

Note that the actual perform() call happens in the main useTask() function.
This is because otherwise you can call the task
repeatedly (synchronously) while the dispatch will fire asynchronously and
be out of date.